+++ backend/controllers/jobController.js
@@ -93,6 +93,9 @@ export const updateJob = catchAsyncErrors(async (req, res, next) => {
   if (!job) {
     return next(new ErrorHandler("OOPS! Job not found.", 404));
   }
+  if (job.postedBy.toString() !== req.user._id.toString()) {
+    return next(new ErrorHandler("Not authorized to update this job.", 403));
+  }
   job = await Job.findByIdAndUpdate(id, req.body, {
     new: true,
     runValidators: true,
@@ -115,6 +118,9 @@ export const deleteJob = catchAsyncErrors(async (req, res, next) => {
   const job = await Job.findById(id);
   if (!job) {
     return next(new ErrorHandler("OOPS! Job not found.", 404));
+  }
+  if (job.postedBy.toString() !== req.user._id.toString()) {
+    return next(new ErrorHandler("Not authorized to delete this job.", 403));
   }
   await job.deleteOne();
   res.status(200).json({
